change register endpoint
