+++ src/middlewares/auth.js
@@ -1,6 +1,5 @@
 const { HttpError, getPayloadAccessToken } = require("../helpers");
 const { modelUser } = require("../models/user");
-const usersServices = require("../service/users");
 
 const auth = async (req, res, next) => {
   const { authorization = "" } = req.headers;
@@ -23,19 +22,4 @@ const auth = async (req, res, next) => {
   }
 };
 
-const isEmailUnique = async (req, res, next) => {
-  try {
-    const { email } = req.body;
-    const user = await usersServices.findUser({ email }, false);
-
-    if (user) {
-      return next(HttpError(409, "Email already in use"));
-    }
-
-    next();
-  } catch (e) {
-    next(e);
-  }
-};
-
-module.exports = { auth, isEmailUnique };
+module.exports = auth;
