+++ src/middlewares/index.js
@@ -3,5 +3,6 @@ const sendMail = require("./sendMail");
 const validateBody = require("./validateBody");
 const isValidId = require("./isValidId");
 const uploadCloud = require("./uploadCloud");
+const isEmailUnique = require("./isEmailUnique");
 
-module.exports = { auth, sendMail, validateBody, isValidId, uploadCloud };
+module.exports = { auth, sendMail, validateBody, isValidId, uploadCloud, isEmailUnique };
